Move blueprint related utils to utils/blueprint
